fix xy0 and offsets in shape and moments
HSM needs the position in the coordinate system of the image it's given
(without any XY0 that it's not told about). There are a few different
offsets involved:
* The image's XY0; this may or may not be related to how things are laid
  out in memory, and needs to be taken into account.
* The offset to the footprint bounding box; this offset is already taken
  into account in the image given to HSM.
* The unit offset required to convert between LSST code and FITS/HSM.

I had confused the first two, but the tests passed. Extending the tests
to move the source away from the lower-left corner of the image and adding
an XY0 revealed the problem.