remove conversion between FITS and LSST LLC conventions
It is not HSM that uses a 1,1 lower-left corner, but GalSim.
However, we can specify a 0,0 lower-left corner if we desire, and
then we can remove the addition and subtraction of 1 in the positions.

The subtraction of 1 in the test of the centroid against the value
out of GalSim needs to stay --- it's coming out of GalSim, not HSM,
so has a 1,1 lower-left corner.