fix variance statistics, used for shape.hsm.<method>.sigma
We measure the median variance around the object, and pass that to
HSM for its error calculation. However, the variance image and mask
image passed to afw::math::makeStatistics were of different sizes,
meaning the mask was not correctly applied. This produced NAN errors
for all sources when pixels were all masked at the lower-left of the
full-size mask image. Fixed by providing the appropriate mask for
the variance image.